git-istage: patch gitconfig: user credentials tmux: socket control, config

+++ bin/tmux_session.sh
@@ -2,16 +2,20 @@
 
 SOCKETS_DIR=~/.tmux_sockets
 
+killall -10 -q tmux
+
 if [ "$#" -lt 1 ]; then
 		ls "$SOCKETS_DIR"
 		exit
 fi
 
 if [ -e "$SOCKETS_DIR/$1" ]; then
-		tmux -S "$SOCKETS_DIR/$1" attach
+		if ! tmux -S "$SOCKETS_DIR/$1" attach
+		then
+				rm "$SOCKETS_DIR/$1"
+				tmux -S "$SOCKETS_DIR/$1" new-session -s "$1"
+		fi
 else
 		tmux -S "$SOCKETS_DIR/$1" new-session -s "$1"
 fi
 
-rm "$SOCKETS_DIR/$1"
-killall -10 tmux
